split api and discovery services
